requirements.txt
ariadne==0.13.0
click==8.0.3
Flask==2.0.2
graphql-core==3.1.6
grpcio==1.41.0
grpcio-tools==1.41.0
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
protobuf==3.18.1
python-dotenv==0.19.1
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.2
Werkzeug==2.0.2
